map structure improvement
